app/requirements.txt
email-validator==2.1.0.post1
fastapi==0.109.2
httptools==0.6.1
httpx==0.26.0
itsdangerous==2.1.2
Jinja2==3.1.3
orjson==3.9.13
pipenv==2023.12.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.9
PyYAML==6.0.1
SQLAlchemy==2.0.26
ujson==5.9.0
unicorn==2.0.1.post1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
psycopg2-binary
passlib
PyJWT
bcrypt
python-jose
pg8000==1.19.5


